Move websocket related tests to correct directory
