+++ backend/tests/websockets/connection_manager_test.py
